requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
psycopg2==2.9.9
flask-wtf==1.2.1
wtforms==3.1.2
flask-bcrypt==1.0.1
werkzeug==3.0.3